NOMA - Brain Power lyrics
Are you ready?
Are you ready?

Adrenaline is pumping
Adrenaline is pumping
Generator. Automatic lover
Atomic. Atomic. Overdrive
Blockbuster. Brain power
Call me a leader. Cocaine
Don't you try it, don't you try it
Innovator. Killer machine
There's no fate. Take control
Brain power

Let the bass kick
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A- JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A-
JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A-
JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA-O----------

Adrenaline is pumping
Adrenaline is pumping
Generator. Automatic lover
Atomic. Atomic. Overdrive
Blockbuster. Brain power
Call me a leader. Cocaine
Don't you try it, don't you try it
Innovator. Killer machine
There's no fate. Take control
Brain power

Let the bass kick
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A- JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A-
JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A-
JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA-O----------

Let the bass kick
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A- JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A-
JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA
O-oooooooooo AAAAE-A-A-I-A-U- JO-oooooooooooo AAE-O-A-A-U-U-A- E-eee-ee-eee AAAAE-A-E-I-E-A-
JO-ooo-oo-oo-oo EEEEO-A-AAA-AAAA-O----------